14811: Find a better OOP approach.
