+++ src/xcode/ENA/ENA/Source/AppDelegate & Globals/Globals.swift
@@ -19,32 +19,6 @@ var isScreenshotMode: Bool {
 
 #endif
 
-/// Determines the CWA Hibernation State, beginning on 01.06.2023.
-var isHibernationState: Bool {
-	#if RELEASE
-
-	var hibernationStartDateComponents = DateComponents()
-	hibernationStartDateComponents.year = 2023
-	hibernationStartDateComponents.month = 6
-	hibernationStartDateComponents.day = 1
-	hibernationStartDateComponents.hour = 0
-	hibernationStartDateComponents.minute = 0
-	hibernationStartDateComponents.second = 0
-	hibernationStartDateComponents.timeZone = .utcTimeZone
-	
-	guard let hibernationStartDate = Calendar.current.date(from: hibernationStartDateComponents) else {
-		fatalError("The hibernation start date couldn't be created.")
-	}
-	
-	return Date() >= hibernationStartDate
-
-	#else
-
-	return true // to.do Dev Menu setting, see https://jira-ibs.wbs.net.sap/browse/EXPOSUREAPP-14812
-	
-	#endif
-}
-
 /// Unconditionally logs, prints a given message and stops execution.
 ///
 /// This 'catches' fatal errors by logging them first as errors… and then failing gloriously.
